requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
Django==4.1
django-crispy-forms==1.14.0
image==1.5.33
Pillow==9.2.0
python-decouple==3.6
six==1.16.0
sqlparse==0.4.2
